CRs              SHA_ID     Commit Message
----------------------------------------------------------------------
4212963           Idd30ca9d sdm: Disable dynamic dest scalar for CWB idle fallback
4249089           If665d108 display-core: update ABC luminance configuration in TVM
4249516           I09d1ea95 display-core: fix clang issues
4246232           I3647db0e snapalloc: Include VENUS format.
4247003           I6ab95200 display-core: config: fix marking
4241780           Ic5bb757c dal: add conditional logging for cursor pipe
4234304           If7754cdc display-core: add soc id for new Canoe variants
4096610           I1e613ee1 sdm: Wait synchronously on retire fence during power-on
4145443           I4e16c3e3 sde-drm: Add lock to synchronize panel prop operations in null commit
4222482           Ib59695c7 sdmclient: Skip redundant SDMServices deinit/init
4238851           I8268c444 sdm: Add DCI-P3L CSC support
4224292           I14fabab2 display-core: add sRGB mode in WCG supported virtual display init
4219316           I45be6df6 sdm: Support Dpu Dma mode
4238939           I609564ff snapalloc: Make BufferType metadata Settable

CRs-Included: 4222482,4249516,4247003,4238939,4212963,4145443,4224292,4249089,4238851,4234304,4246232,4219316,4241780,4096610 .

+++ sdm/libs/core/color_manager.cpp
@@ -1881,7 +1881,7 @@ bool DPUColorManager::IsValidateNeeded() {
   for (int i = 1; i < color_mgr_cnt; i++) {
     if (needed[0] != needed[i]) {
       DLOGW("Need validate for DPU's are different, DPU0=%d, DPU%d=%d",
-                                    needed[0], i, needed[i]);
+                                    (int)needed[0], i, (int)needed[i]);
     }
   }
